import stdlib module from puppet module forge
The Jenkins Puppet module requires the stdlib module. Import it into our
tree here. We'll track any changes as separate commits.

Since we're still on Puppet version 2.6 from EPEL, we'll use an older
version of the stdlib module that is compatible with Puppet 2.6.

This code is licensed under the ASL 2.0.

Code downloaded from
http://forge.puppetlabs.com/puppetlabs/stdlib/2.6.0.tar.gz