add gitorious.org modules
Add in all the modules and configurations necessary to operate Jenkins
for Gitorious.org. Jenkins builds are now working (I tested mainline and
dolt).

The name of the host is currently set to "gitoriousci.ktdreyer.com", and
I anticipate that we will change this to something.gitorious.org in the
future.

Adjust the "lint" task in the Makefile by skipping the 80-character
check. The SSH pubkey for the slave(s) is longer than 80 characters.