add an LXC guest with ruby193
Create a new "hypervisor" module and add this to the main
ci.gitorious.org server. This installs libvirtd and manages the
appropriate configurations and services.

Create a new "centos-6-ruby193" LXC guest to run on this hypervisor.
There are still a couple manual steps (such as running the "install"
shell script), but the process is largely automated and
version-controlled "enough" in Puppet for now.

Add the "centos-6-ruby193.ci.gitorious.org" host to the site manifest.
There's a bit of copying & pasting here, so we could abstract this
further down the road if warranted.

Modify the gitorious_jenkins::builder class to optionally take an "scl"
parameter. Use this parameter to install the ruby193 packages from Red
Hat's Software Collections project.